Completed implementing most of interfaces. Also, seperated the interface implementations into various files using a partial class
