+++ OutrageousNumbers/OutrageousInt.Identity.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Numerics;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace OutrageousNumbers
+{
+    public partial class OutrageousInt
+    {
+        public static OutrageousInt AdditiveIdentity => OutrageousInt.Zero;
+
+        public static OutrageousInt MultiplicativeIdentity => OutrageousInt.One;
+    }
+}
